+++ CHANGELOG.md
@@ -11,7 +11,8 @@
 - [Deprecate --skip-schedule and replace by --hibernate-schedule-off and --hibernate-schedule-on in cm hibernate/run cc #209](https://github.com/stolostron/cm-cli/issues/209)
 - [Add hibernate schedule option in cm set cc #212](https://github.com/stolostron/cm-cli/issues/212)
 - Removed `--skip-schedule` from `cm create cc` as it wa ineffective. Use `cm set cc [--hibernate-schedule-on|--hibernate-schedule-off]` to set it.
-
+- [add a cm to delete an hypershift deployment #215](https://github.com/stolostron/cm-cli/issues/215)
+- [Add cm to create an hypershift deployment on a given hostingcluster #214](https://github.com/stolostron/cm-cli/issues/214)
 ## Breaking changes
 
 ## Bug fixes